修复bug
请叫我bug狂魔

+++ DanDanPlayForMac/ViewModel/PlayViewModel.h
@@ -88,7 +88,7 @@
  *
  *  @param index 下标
  */
-- (void)reloadDanmakuWithIndex:(NSInteger)index completionHandler:(void(^)(NSInteger progress, NSString *videoMatchName, NSError *error))complete;
+- (void)reloadDanmakuWithIndex:(NSInteger)index completionHandler:(void(^)(CGFloat progress, NSString *videoMatchName, NSError *error))complete;
 /**
  *  初始化
  *
